WIP Refactor Overview Detail
