+++ src/components/Nav/Page/RenderComponentScroll.tsx
@@ -5,6 +5,7 @@ const TOP_PADDING = 76 + 118;
 
 interface Props {
   className?: any;
+  onResize?: (height: number) => void;
 }
 
 interface State {
@@ -27,7 +28,16 @@ export class RenderComponentScroll extends React.Component<Props, State> {
   }
 
   updateWindowDimensions = () => {
-    this.setState({ height: window.innerHeight - TOP_PADDING });
+    this.setState(
+      {
+        height: window.innerHeight - TOP_PADDING
+      },
+      () => {
+        if (this.props.onResize) {
+          this.props.onResize(this.state.height);
+        }
+      }
+    );
   };
 
   render() {
