+++ src/pages/AppDetails/AppInfo/AppDescription.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { DisplayMode, HealthIndicator } from '../../../components/Health/HealthIndicator';
 import MissingSidecar from '../../../components/MissingSidecar/MissingSidecar';
 import { AppHealth } from '../../../types/Health';
 import { App, AppWorkload } from '../../../types/App';
@@ -13,23 +12,18 @@ import {
   DataListItem,
   DataListItemCells,
   DataListItemRow,
-  Grid,
-  GridItem,
   List,
   ListItem,
-  PopoverPosition,
-  Stack,
-  StackItem,
   Title
 } from '@patternfly/react-core';
 import GraphDataSource from '../../../services/GraphDataSource';
-import MiniGraphCard from '../../../components/CytoscapeGraph/MiniGraphCard';
 import './AppDescription.css';
 import { style } from 'typestyle';
 
 const iconStyle = style({
   margin: '0 16px 0 0',
-  padding: '24px 0 0 0'
+  padding: '0 0 0 0',
+  display: 'inline-block'
 });
 
 const resourceListStyle = style({
@@ -58,10 +52,15 @@ class AppDescription extends React.Component<AppDescriptionProps> {
   private renderWorkloadItem(namespace: string, workload: AppWorkload) {
     return (
       <ListItem key={`AppWorkload_${workload.workloadName}`}>
-        <Link to={this.workloadLink(namespace, workload.workloadName)}>{workload.workloadName}</Link>
-        {!workload.istioSidecar && (
-          <MissingSidecar namespace={namespace} tooltip={true} style={{ marginLeft: '10px' }} text={''} />
-        )}
+        <span>
+          <div key="workload-icon" className={iconStyle}>
+            <Badge>W</Badge>
+          </div>
+          <Link to={this.workloadLink(namespace, workload.workloadName)}>{workload.workloadName}</Link>
+          {!workload.istioSidecar && (
+            <MissingSidecar namespace={namespace} tooltip={true} style={{ marginLeft: '10px' }} text={''} />
+          )}
+        </span>
       </ListItem>
     );
   }
@@ -69,7 +68,12 @@ class AppDescription extends React.Component<AppDescriptionProps> {
   private renderServiceItem(namespace: string, _appName: string, serviceName: string) {
     return (
       <ListItem key={`AppService_${serviceName}`}>
-        <Link to={this.serviceLink(namespace, serviceName)}>{serviceName}</Link>
+        <div key="service-icon" className={iconStyle}>
+          <Badge>S</Badge>
+        </div>
+        <span>
+          <Link to={this.serviceLink(namespace, serviceName)}>{serviceName}</Link>
+        </span>
       </ListItem>
     );
   }
@@ -86,9 +90,6 @@ class AppDescription extends React.Component<AppDescriptionProps> {
       workloads.length > 0 ? workloads.map(wkd => this.renderWorkloadItem(ns, wkd)) : this.renderEmptyItem('workloads');
 
     return [
-      <div key="workload-icon" className={iconStyle}>
-        <Badge>W</Badge>
-      </div>,
       <div key="workload-list" className={resourceListStyle}>
         <Title headingLevel="h3" size="lg" className={titleStyle}>
           Workloads
@@ -107,9 +108,6 @@ class AppDescription extends React.Component<AppDescriptionProps> {
         : this.renderEmptyItem('services');
 
     return [
-      <div key="service-icon" className={iconStyle}>
-        <Badge>S</Badge>
-      </div>,
       <div key="service-list" className={resourceListStyle}>
         <Title headingLevel="h3" size="lg" className={titleStyle}>
           Services
@@ -122,60 +120,28 @@ class AppDescription extends React.Component<AppDescriptionProps> {
   render() {
     const app = this.props.app;
     return app ? (
-      <Grid gutter="md">
-        <GridItem span={4}>
-          <Card style={{ height: '100%' }}>
-            <CardHeader>
-              <Title headingLevel="h3" size="2xl">
-                {' '}
-                Application Structure
-              </Title>
-            </CardHeader>
-            <CardBody className="noPadding">
-              <DataList aria-label="workloads and services">
-                <DataListItem aria-labelledby="Workloads">
-                  <DataListItemRow>
-                    <DataListItemCells dataListCells={this.workloadList()} />
-                  </DataListItemRow>
-                </DataListItem>
-                <DataListItem aria-labelledby="Services">
-                  <DataListItemRow>
-                    <DataListItemCells dataListCells={this.serviceList()} />
-                  </DataListItemRow>
-                </DataListItem>
-              </DataList>
-            </CardBody>
-          </Card>
-        </GridItem>
-        <GridItem span={4}>
-          <MiniGraphCard dataSource={this.props.miniGraphDataSource} />
-        </GridItem>
-        <GridItem span={4}>
-          <Card style={{ height: '100%' }}>
-            <CardHeader>
-              <Title headingLevel="h3" size="2xl">
-                {' '}
-                Health Overview{' '}
-              </Title>
-            </CardHeader>
-            <CardBody>
-              <Stack>
-                <StackItem id="health" className={'stack_service_details'}>
-                  <Title headingLevel="h3" size="lg" className={titleStyle}>
-                    Overall Health
-                  </Title>
-                  <HealthIndicator
-                    id={app.name}
-                    health={this.props.health}
-                    mode={DisplayMode.LARGE}
-                    tooltipPlacement={PopoverPosition.left}
-                  />
-                </StackItem>
-              </Stack>
-            </CardBody>
-          </Card>
-        </GridItem>
-      </Grid>
+      <Card>
+        <CardHeader>
+          <Title headingLevel="h3" size="2xl">
+            {' '}
+            Application
+          </Title>
+        </CardHeader>
+        <CardBody className="noPadding">
+          <DataList aria-label="workloads and services">
+            <DataListItem aria-labelledby="Workloads">
+              <DataListItemRow>
+                <DataListItemCells dataListCells={this.workloadList()} />
+              </DataListItemRow>
+            </DataListItem>
+            <DataListItem aria-labelledby="Services">
+              <DataListItemRow>
+                <DataListItemCells dataListCells={this.serviceList()} />
+              </DataListItemRow>
+            </DataListItem>
+          </DataList>
+        </CardBody>
+      </Card>
     ) : (
       'Loading'
     );
