+++ src/pages/ServiceDetails/ServiceInfo.tsx
@@ -1,23 +1,23 @@
 import * as React from 'react';
 import { style } from 'typestyle';
-import { Grid, GridItem, Tab } from '@patternfly/react-core';
+import { Grid, GridItem } from '@patternfly/react-core';
 import ServiceId from '../../types/ServiceId';
-import ServiceInfoDescription from './ServiceInfo/ServiceInfoDescription';
-import { ServiceDetailsInfo, validationToSeverity } from '../../types/ServiceInfo';
-import ServiceInfoWorkload from './ServiceInfo/ServiceInfoWorkload';
-import { ObjectValidation, PeerAuthentication, Validations, ValidationTypes } from '../../types/IstioObjects';
-import ParameterizedTabs, { activeTab } from '../../components/Tab/Tabs';
-import ErrorBoundaryWithMessage from '../../components/ErrorBoundary/ErrorBoundaryWithMessage';
-import Validation from '../../components/Validations/Validation';
+import ServiceDescription from './ServiceInfo/ServiceDescription';
+import { ServiceDetailsInfo } from '../../types/ServiceInfo';
+import { ObjectValidation, PeerAuthentication, Validations } from '../../types/IstioObjects';
+import { activeTab } from '../../components/Tab/Tabs';
 import { RenderComponentScroll } from '../../components/Nav/Page';
 import { PromisesRegistry } from 'utils/CancelablePromises';
 import { DurationInSeconds, TimeInMilliseconds } from 'types/Common';
 import GraphDataSource from 'services/GraphDataSource';
-import IstioConfigSubList from '../../components/IstioConfigSubList/IstioConfigSubList';
 import { drToIstioItems, vsToIstioItems } from '../../types/IstioConfigList';
 import { KialiAppState } from '../../store/Store';
 import { connect } from 'react-redux';
 import { durationSelector } from '../../store/Selectors';
+import MiniGraphCard from '../../components/CytoscapeGraph/MiniGraphCard';
+import ServiceWorkload from './ServiceInfo/ServiceWorkload';
+import HealthCard from '../../components/Health/HealthCard';
+import IstioConfigCard from '../../components/IstioConfigCard/IstioConfigCard';
 
 interface Props extends ServiceId {
   duration: DurationInSeconds;
@@ -30,6 +30,7 @@ interface Props extends ServiceId {
 
 type ServiceInfoState = {
   currentTab: string;
+  tabHeight?: number;
 };
 
 interface ValidationChecks {
@@ -37,16 +38,12 @@ interface ValidationChecks {
   hasDestinationRuleChecks: boolean;
 }
 
-const tabIconStyle = style({
-  fontSize: '0.9em'
-});
-
 const tabName = 'list';
 const defaultTab = 'workloads';
-const paramToTab: { [key: string]: number } = {
-  workloads: 0,
-  istioconfig: 1
-};
+
+const fullHeightStyle = style({
+  height: '100%'
+});
 
 class ServiceInfo extends React.Component<Props, ServiceInfoState> {
   private promises = new PromisesRegistry();
@@ -109,10 +106,6 @@ class ServiceInfo extends React.Component<Props, ServiceInfoState> {
     return validationChecks;
   }
 
-  private errorBoundaryMessage(resourceName: string) {
-    return `One of the ${resourceName} associated to this service has an invalid format`;
-  }
-
   private getServiceValidation(): ObjectValidation | undefined {
     if (this.props.validations && this.props.validations.service && this.props.serviceDetails) {
       return this.props.validations.service[this.props.serviceDetails.service.name];
@@ -123,14 +116,14 @@ class ServiceInfo extends React.Component<Props, ServiceInfoState> {
   render() {
     const workloads = this.props.serviceDetails?.workloads || [];
     const validationChecks = this.validationChecks();
-    const getSeverityIcon: any = (severity: ValidationTypes = ValidationTypes.Error) => (
+    /*    const getSeverityIcon: any = (severity: ValidationTypes = ValidationTypes.Error) => (
       <span className={tabIconStyle}>
         {' '}
         <Validation severity={severity} />
       </span>
-    );
+    );*/
 
-    const getValidationIcon = (keys: string[], types: string[]) => {
+    /*    const getValidationIcon = (keys: string[], types: string[]) => {
       let severity = ValidationTypes.Warning;
       keys.forEach(key => {
         types.forEach(type => {
@@ -143,25 +136,14 @@ class ServiceInfo extends React.Component<Props, ServiceInfoState> {
         });
       });
       return getSeverityIcon(severity);
-    };
+    };*/
 
-    let istioTabTitle: JSX.Element | undefined;
     if (this.props.serviceDetails) {
-      let istioConfigIcon = undefined;
       if (validationChecks.hasVirtualServiceChecks || validationChecks.hasDestinationRuleChecks) {
         const names: string[] = [];
         this.props.serviceDetails.virtualServices?.items.forEach(vs => names.push(vs.metadata.name));
         this.props.serviceDetails.destinationRules?.items.forEach(dr => names.push(dr.metadata.name));
-        istioConfigIcon = getValidationIcon(names, ['virtualservice', 'destinationrule']);
       }
-      istioTabTitle = (
-        <>
-          Istio Config (
-          {this.props.serviceDetails.virtualServices.items.length +
-            this.props.serviceDetails.destinationRules.items.length}
-          ){istioConfigIcon}
-        </>
-      );
     }
 
     const vsIstioConfigItems = this.props.serviceDetails?.virtualServices
@@ -172,31 +154,14 @@ class ServiceInfo extends React.Component<Props, ServiceInfoState> {
       : [];
     const istioConfigItems = vsIstioConfigItems.concat(drIstioConfigItems);
 
-    return (
-      <>
-        <RenderComponentScroll>
-          <Grid gutter={'md'}>
-            <GridItem span={12}>
-              <ServiceInfoDescription
-                name={this.props.serviceDetails?.service.name || ''}
-                namespace={this.props.namespace}
-                createdAt={this.props.serviceDetails?.service.createdAt || ''}
-                resourceVersion={this.props.serviceDetails?.service.resourceVersion || ''}
-                additionalDetails={this.props.serviceDetails?.additionalDetails || []}
-                istioEnabled={this.props.serviceDetails?.istioSidecar}
-                labels={this.props.serviceDetails?.service.labels}
-                selectors={this.props.serviceDetails?.service.selectors}
-                ports={this.props.serviceDetails?.service.ports}
-                type={this.props.serviceDetails?.service.type}
-                ip={this.props.serviceDetails?.service.ip}
-                endpoints={this.props.serviceDetails?.endpoints}
-                health={this.props.serviceDetails?.health}
-                externalName={this.props.serviceDetails?.service.externalName}
-                validations={this.getServiceValidation()}
-                miniGraphDatasource={this.graphDataSource}
-              />
-            </GridItem>
-            {this.props.serviceDetails && (
+    // RenderComponentScroll handles height to provide an inner scroll combined with tabs
+    // This height needs to be propagated to minigraph to proper resize in height
+    // Graph resizes correctly on width
+    const height = this.state.tabHeight ? this.state.tabHeight - 115 : 300;
+    const graphContainerStyle = style({ width: '100%', height: height });
+
+    /*
+      this.props.serviceDetails && (
               <GridItem span={12}>
                 <ParameterizedTabs
                   id="service-tabs"
@@ -225,6 +190,51 @@ class ServiceInfo extends React.Component<Props, ServiceInfoState> {
                 </ParameterizedTabs>
               </GridItem>
             )}
+     */
+
+    return (
+      <>
+        <RenderComponentScroll onResize={height => this.setState({ tabHeight: height })}>
+          <Grid gutter={'md'} className={fullHeightStyle}>
+            <GridItem span={6} rowSpan={2}>
+              <MiniGraphCard dataSource={this.graphDataSource} graphContainerStyle={graphContainerStyle} />
+            </GridItem>
+            <GridItem span={3}>
+              <ServiceDescription
+                name={this.props.serviceDetails?.service.name || ''}
+                namespace={this.props.namespace}
+                createdAt={this.props.serviceDetails?.service.createdAt || ''}
+                resourceVersion={this.props.serviceDetails?.service.resourceVersion || ''}
+                additionalDetails={this.props.serviceDetails?.additionalDetails || []}
+                istioEnabled={this.props.serviceDetails?.istioSidecar}
+                labels={this.props.serviceDetails?.service.labels}
+                selectors={this.props.serviceDetails?.service.selectors}
+                ports={this.props.serviceDetails?.service.ports}
+                type={this.props.serviceDetails?.service.type}
+                ip={this.props.serviceDetails?.service.ip}
+                endpoints={this.props.serviceDetails?.endpoints}
+                externalName={this.props.serviceDetails?.service.externalName}
+                validations={this.getServiceValidation()}
+              />
+            </GridItem>
+            <GridItem span={3}>
+              {this.props.serviceDetails ? (
+                <HealthCard name={this.props.serviceDetails.service.name} health={this.props.serviceDetails.health} />
+              ) : (
+                'Loading'
+              )}
+            </GridItem>
+            <GridItem span={3}>
+              <ServiceWorkload
+                namespace={this.props.namespace}
+                service={this.props.service}
+                istioSidecar={this.props.serviceDetails ? this.props.serviceDetails.istioSidecar : false}
+                workloads={workloads}
+              />
+            </GridItem>
+            <GridItem span={3}>
+              <IstioConfigCard name={this.props.service} items={istioConfigItems} />
+            </GridItem>
           </Grid>
         </RenderComponentScroll>
       </>
