+++ src/pages/ServiceDetails/ServiceInfo/__tests__/ServiceInfoWorkloads.test.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
-import ServiceInfoWorkload from '../ServiceInfoWorkload';
+import ServiceWorkload from '../ServiceInfoWorkload';
 import { shallowToJson } from 'enzyme-to-json';
 import { Service } from './ServiceMock';
 
 describe('#ServiceInfoWorkload render correctly with data', () => {
   it('should render service pods', () => {
-    const wrapper = shallow(<ServiceInfoWorkload service={Service} workloads={Service.workloads} namespace={'ns'} />);
+    const wrapper = shallow(<ServiceWorkload service={Service} workloads={Service.workloads} namespace={'ns'} />);
     expect(shallowToJson(wrapper)).toBeDefined();
     expect(shallowToJson(wrapper)).toMatchSnapshot();
   });
